requirements.txt
alpaca-trade-api==3.1.1
pandas==2.2.2
numpy>=1.24.3
yfinance==0.2.28
ta==0.10.2
python-dotenv==1.0.0
jupyter==1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0
requests>=2.31.0
